week-4/assignment-2/submissions/rohannarasayya/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
adjustText>=0.8.0
